remove symlinks
